% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_matrix.R
\name{add_matrix}
\alias{add_matrix}
\title{add_matrix}
\usage{
add_matrix(call., formula, nonlinear = FALSE, envir = parent.frame(), ...)
}
\arguments{
\item{call.}{a \emph{pseudo} \link[mgcv]{gam} \link[base]{call}}

\item{formula}{..}

\item{nonlinear}{..}

\item{envir}{..}

\item{...}{..}
}
\description{
add_matrix
}
\examples{
# see ?`gam.matrix-package`

}
