% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{cor_xy}
\alias{cor_xy}
\alias{cor_xy.gam_matrix}
\title{\eqn{xy}-Correlation}
\usage{
cor_xy(object, ...)

\method{cor_xy}{gam_matrix}(object, probs = 0.5, ...)
}
\arguments{
\item{object}{a \link{gam_matrix} object}

\item{...}{additional parameters, currently not in use}

\item{probs}{\link[base]{numeric} scalar or \link[base]{vector} \eqn{\tilde{p}},
taking values between 0 and 1, see function \link[stats]{quantile}.}
}
\value{
Function \link{cor_xy.gam_matrix} returns a \link[base]{numeric} scalar or \link[base]{vector} of
\link[stats]{cor}relation(s).
}
\description{
\eqn{xy}-Correlation
}
