% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{gam.matrix-package}
\alias{gam.matrix}
\alias{gam.matrix-package}
\title{\link[mgcv]{gam} with one-and-only-one \link[base]{matrix} Predictor}
\description{
Function \link{gam_matrix} ..

Function \link{cor_xy.gam} ..

Function \link{sign_adjust.gam_matrix} ..

Function \link{predict.gam_matrix} ..

Functions \link{persp.gam_matrix} and \link{contour.gam_matrix} ..

Function \link{integrandSurface} ..
}
\examples{
#library(mgcv)
library(survival)
library(spatstat.grouped)
data(Ki67, package = 'spatstat.grouped.data')
Ki67 = within.data.frame(Ki67, expr = {
 Ki67 = log1p(Ki67)
 PFS = Surv(time = recfreesurv_mon, event = recurrence)
})
#(npt = length(unique(Ki67$patientID))) # 622
Ki67q = grouped_ppp(Ki67 ~ PFS + node + Tstage | patientID/tissueID, data = Ki67) |>
  aggregate_quantile(by = ~ patientID, probs = seq.int(from = .01, to = .99, by = .01))
  
#set.seed(234); id = sort.int(sample.int(n = npt, size = 500L))
#Ki67q_0 = Ki67q[id, , drop = FALSE] # training set
#Ki67q_1 = Ki67q[-id, , drop = FALSE] # test set

m0 = gam_matrix(PFS ~ Ki67.quantile, data = Ki67q)

m0a = quote(gam(formula = PFS[,1L] ~ 1, data = Ki67q, family = cox.ph(), weight = PFS[,2L])) |>
 add_matrix(~ Ki67.quantile)
#ThomasJeffersonUniv::relaxed_identical(m0, m0a)

m0b = quote(gam(formula = PFS ~ 1, data = Ki67q, family = cox.ph())) |>
 add_matrix(~ Ki67.quantile)
#ThomasJeffersonUniv::relaxed_identical(m0, m0b)

#fr = sign_adjust(m0)
#predict(m0, newdata = Ki67q_1)
m0 |> predict_k_fold.gam_matrix(k = 10L)
persp(m0)
\donttest{
# integrandSurface(gam_matrix(PFS ~ Ki67.quantile, data = Ki67q_0)) # correct
integrandSurface(m0) # bug with proj_beta!!!!
#integrandSurface(m0, newdata = Ki67q_1)
} # to save time

m1 = gam_matrix(PFS ~ Ki67.quantile, data = Ki67q, nonlinear = TRUE)
m1b = quote(gam(formula = PFS ~ 1, data = Ki67q, family = cox.ph())) |>
 add_matrix(~ Ki67.quantile, nonlinear = TRUE)
#which(!mapply(ThomasJeffersonUniv::relaxed_identical, m1, m1b))

m1 |> predict_k_fold.gam_matrix(k = 10L)
\donttest{
integrandSurface(m1)
#integrandSurface(m1, newdata = Ki67q_1)
} # to save time

\donttest{integrandSurface(m0, m1)}

}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

Authors:
\itemize{
  \item Inna Chervoneva \email{Inna.Chervoneva@jefferson.edu} (\href{https://orcid.org/0000-0002-9104-4505}{ORCID})
}

}
