% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_matrix.R
\name{gam_matrix}
\alias{gam_matrix}
\title{\link[mgcv]{gam} with one-and-only-one \link[base]{matrix} predictor}
\usage{
gam_matrix(formula, data, family, fit = TRUE, nonlinear = FALSE, ...)
}
\arguments{
\item{formula}{\link[stats]{formula}, e.g., \code{y~X}.
Response \eqn{y} may be \link[base]{double}, \link[base]{logical} and \link[survival]{Surv}.
Functional predictor \eqn{X} is a \link[base]{double} \link[base]{matrix}.
The \link[base]{colnames} of \eqn{X} must be convertible to \link[base]{numeric} \link[base]{vector},
indicating the \emph{common tabulating grid} shared by all subjects.}

\item{data}{\link[base]{data.frame}}

\item{family}{\link[stats]{family} object,
see function \link[mgcv]{gam}.
Default values are
\itemize{
\item \code{mgcv::cox.ph()} for \link[survival]{Surv} response \eqn{y};
\item \code{binomial(link = 'logit')} for \link[base]{logical} response \eqn{y};
\item \code{gaussian(link = 'identity')} for \link[base]{double} response \eqn{y}
}}

\item{fit}{see function \link[mgcv]{gam}}

\item{nonlinear}{\link[base]{logical} scalar,
whether to use nonlinear or linear functional model.
Default \code{FALSE}}

\item{...}{additional parameters for functions \link[mgcv]{s} and \link[mgcv]{ti},
most importantly \code{k}}
}
\value{
Function \link{gam_matrix} returns a \link{gam_matrix} object, which \link[base]{inherits} from class \link[mgcv]{gam}.
}
\description{
\link[mgcv]{gam} with one-and-only-one \link[base]{matrix} predictor
}
\examples{
# see ?`gam.matrix-package`
}
