% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_k_fold.R
\name{predict_k_fold.gam_matrix}
\alias{predict_k_fold.gam_matrix}
\title{Predict via Cross Validation}
\usage{
predict_k_fold.gam_matrix(object, k, ...)
}
\arguments{
\item{object}{a \link{gam_matrix} object}

\item{k}{\link[base]{integer} scalar}

\item{...}{additional parameters, currently not in use}
}
\description{
Predict via Cross Validation
}
